cypress testing has been installed and configured
