+++ cypress/e2e/genreSelect_spec.cy.ts
@@ -1,0 +1,5 @@
+describe('genre select spec', () => {
+  it('passes', () => {
+    cy.visit('http://localhost:3000')
+  })
+})
