+++ package.json
@@ -46,7 +46,7 @@
     "cypress": "^12.9.0",
     "jest": "^29.5.0",
     "ts-node": "^10.9.1",
-    "typescript": "^5.0.2",
+    "typescript": "^5.0.4",
     "wdio-chromedriver-service": "^8.1.1"
   }
 }
