+++ src/components/Search/__test__/SearchFrom.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render, fireEvent } from '@testing-library/react';
-import { SearchForm } from './SearchForm';
+import { SearchForm } from '../SearchForm';
 
 describe('SearchForm', () => {
   it('should render without throwing an error', () => {
